requirements.txt
duckdb
rdkit
pandas
numpy
google-cloud-bigquery
